requirements.txt
python-rpi.gpio
python3-rpi.gpio
pygame
